+++ test/Middleware/ImplicitHeadMiddlewareTest.php
@@ -83,7 +83,8 @@ class ImplicitHeadMiddlewareTest extends TestCase
     public function testReturnsResultOfHandlerWhenRouteResultDoesNotComposeRoute()
     {
         $result = $this->prophesize(RouteResult::class);
-        $result->getMatchedRoute()->willReturn(null);
+        $result->getAllowedMethods()->willReturn([]);
+        $result->getMatchedRoute()->willReturn(false);
 
         $request = $this->prophesize(ServerRequestInterface::class);
         $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
@@ -106,6 +107,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $route->implicitHead()->willReturn(false);
 
         $result = $this->prophesize(RouteResult::class);
+        $result->getAllowedMethods()->willReturn([RequestMethod::METHOD_HEAD]);
         $result->getMatchedRoute()->will([$route, 'reveal']);
 
         $request = $this->prophesize(ServerRequestInterface::class);
@@ -127,9 +129,9 @@ class ImplicitHeadMiddlewareTest extends TestCase
     {
         $route = $this->prophesize(Route::class);
         $route->implicitHead()->willReturn(true);
-        $route->allowsMethod(RequestMethod::METHOD_GET)->willReturn(false);
 
         $result = $this->prophesize(RouteResult::class);
+        $result->getAllowedMethods()->willReturn([RequestMethod::METHOD_HEAD]);
         $result->getMatchedRoute()->will([$route, 'reveal']);
 
         $request = $this->prophesize(ServerRequestInterface::class);
@@ -148,9 +150,9 @@ class ImplicitHeadMiddlewareTest extends TestCase
     {
         $route = $this->prophesize(Route::class);
         $route->implicitHead()->willReturn(true);
-        $route->allowsMethod(RequestMethod::METHOD_GET)->willReturn(true);
 
         $result = $this->prophesize(RouteResult::class);
+        $result->getAllowedMethods()->willReturn([RequestMethod::METHOD_HEAD, RequestMethod::METHOD_GET]);
         $result->getMatchedRoute()->will([$route, 'reveal']);
 
         $request = $this->prophesize(ServerRequestInterface::class);